+++ src/main/java/com/sideproject/withpt/application/gym/service/GymQueryService.java
@@ -4,6 +4,7 @@ import com.sideproject.withpt.application.gym.controller.response.TrainerAllGyms
 import com.sideproject.withpt.application.gym.controller.response.TrainerAllGymsResponse.GymResponse;
 import com.sideproject.withpt.application.gym.exception.GymException;
 import com.sideproject.withpt.application.gym.repositoy.GymQueryRepository;
+import com.sideproject.withpt.application.lesson.controller.response.SearchMemberResponse;
 import com.sideproject.withpt.application.trainer.service.TrainerService;
 import com.sideproject.withpt.domain.gym.Gym;
 import com.sideproject.withpt.domain.gym.GymTrainer;
@@ -24,6 +25,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class GymQueryService {
 
     private final TrainerService trainerService;
+    private final GymService gymService;
     private final GymQueryRepository gymQueryRepository;
 
     public TrainerAllGymsResponse listOfTrainerAllGyms(Long trainerId) {
@@ -44,5 +46,13 @@ public class GymQueryService {
     public Slice<Gym> listOfAllGymsByPageable(Long trainerId, Pageable pageable) {
         Trainer trainer = trainerService.getTrainerById(trainerId);
         return gymQueryRepository.findAllTrainerGymsByPageable(trainer, pageable);
+    }
+
+    public Slice<SearchMemberResponse> searchMembersByGymIdAndName(Long gymId, Long trainerId, String name, Pageable pageable) {
+        Trainer trainer = trainerService.getTrainerById(trainerId);
+        Gym gym = gymService.getGymById(gymId);
+
+        // TODO : 예외 처리 추가
+        return gymQueryRepository.findAllMembersByGymIdAndName(trainer, gym, name, pageable);
     }
 }
